// src/app/models/tabla-desayuno.ts
export interface TablaDesayuno {
    ubicacion: string;
    diagnostico: string;
    nombre: string;
    apellido: string;
    alergias: string;
    dietaIndicada: string;
    dietaAdecuada: string;
    definir: string;
    gustosSi: string;
    gustosNo: string;
    anamnesis: string;
    validado: boolean;
    liquidos: string;
    panificados: string;
    reposteria: string;
    untables: string;
    extras: string;
    otrosEspeciales: string;
  }
  